# api-gateway/docs/swagger.yaml
basePath: /api
definitions:
  main.AuthRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  main.CreateOrderRequest:
    properties:
      product_id:
        type: string
      quantity:
        type: integer
    type: object
  main.CreateProductRequest:
    properties:
      description:
        type: string
      name:
        type: string
      price:
        type: integer
    type: object
  main.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  main.LoginResponse:
    properties:
      access_token:
        type: string
    type: object
  main.MessageResponse:
    properties:
      status:
        type: string
    type: object
  main.OrderResponse:
    properties:
      order_id:
        type: integer
      status:
        type: string
    type: object
  main.ProductIDResponse:
    properties:
      id:
        type: string
    type: object
  main.ProductResponse:
    properties:
      description:
        type: string
      id:
        type: string
      name:
        type: string
      price:
        type: integer
    type: object
  main.RegisterResponse:
    properties:
      user_id:
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  description: API Gateway for the Go microservices E-Commerce project.
  title: E-Commerce API Gateway
  version: "1.0"
paths:
  /login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/main.AuthRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.LoginResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      summary: Login
      tags:
      - Auth
  /logout:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.MessageResponse'
      security:
      - BearerAuth: []
      summary: Logout
      tags:
      - Auth
  /orders:
    post:
      consumes:
      - application/json
      parameters:
      - description: Order Details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/main.CreateOrderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.OrderResponse'
      security:
      - BearerAuth: []
      summary: Create an order
      tags:
      - Orders
  /products:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/main.ProductResponse'
            type: array
      summary: Get all products
      tags:
      - Products
    post:
      consumes:
      - application/json
      parameters:
      - description: Product INFO
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/main.CreateProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.ProductIDResponse'
      security:
      - BearerAuth: []
      summary: Add a new product
      tags:
      - Products
  /register:
    post:
      consumes:
      - application/json
      parameters:
      - description: Credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/main.AuthRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.RegisterResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      summary: Registration
      tags:
      - Auth
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"